Consider per-queue demands when reporting scheduler load.
If a job has been submitted with a hard queue list (which may be a regex)
then we tally that job against the first matching autoscaling queue.

If it doesn't match an autoscaling queue or has no hard list specified we tally
it into the "general" bucket as before. This also applies if we have no
autoscaling groups in the first place (identified by "FlightComputeGroup" in
the queue name).