Do a hostwise analysis for empty nodes.
The previous implementation only considered jobs in the global `bynode.q`
and not individual autoscaling-group queues. But, with `qhost -j` we can get a
list of all SGE compute hosts and the jobs, if any, currently running on them.

The autoscaler `_scale-in` function filters out hosts that aren't part of
autoscaling groups, so we don't need to worry about that here.