Install s3cmd dependency
Rather than attempting to work around lack of this.

Previously if s3cmd was run without passing the `--no-mime-magic`
option, and a file upload was attempted, then the following warning
would be given:

```
WARNING: Module python-magic is not available. Guessing MIME types based on file extensions.
```

We attempt to work around this in some places but rather than do this,
and do this inconsistently, this commit just installs the dependency;
this gets rid of the warning and should mean MIME types are detected
correctly.