Include node and domain in job queues (fixes #36)
This commit includes the login node name and current Flight domain name
when forming the path for job queues in the `alces customize job-queue`
tool; this serves two purposes.

Including the domain fixes
https://github.com/alces-software/clusterware-services/issues/36 by
making clusters with the same name in different domains put and pull job
queue scripts to and from distinct buckets.

Including the node name will allow job queue jobs to be run on any node
in a cluster, once other changes have been made to the
`cluster-customizer/event-periodic` handler hook to allow this, and to
use a different node name here when needed. The node name is currently
hard-coded to that of the login node to preserve the original behavior
of the `alces customize job-queue` tool.