Don't redirect the customizer config to HTTPS.
A complex interaction of firewall rules, SSL and DNS mean that the HTTP->HTTPS
redirect wasn't always working for the customizer. Since we don't really need
this file to be sent over HTTPS anyway, it's easiest to just not redirect.

We do this by enclosing the redirect in a `location /` block, and create
another block for `/customizer` without a redirect.