Unbreak alces-flight-tutorials
Previously, the web-terminal server was running on the same port as the flight
tutorials server.  It would not have been possible to have both installed at
the same time.

Now, the flight tutorials server has been removed and an attempt to use the
web-terminal server has been made.

This should work, but I've decided that this is too risky.  A future commit
will implement something less risky.