Revert changes to alces-flight-tutorials
It no longers tries to make use of alces-web-terminal.  Instead we end
up with two instances of the web-terminal/alces-tutorials-server app
running.

To prevent the two instances of alces-tutorials-server clashing we use
different configuration for the web-terminal instance.

This is much less likely to result in the tutorials being broken in some
way and still leaves a very simple deprecation and removal path for
alces-flight-tutorials service.