Allow personality data to set compute queue config
The `alces compute ...` commands provide tracon with the cluster's
name and network domain.  tracon uses these to determine the
cloudformation stack name.  This does not work for clusters launched
with Flight Launch as the stack name also contains a hash of the user's
email address.

This commit allows Flight Launch to specify the cluster name and domain
that should be used by the `alces compute ...` commands.